AGROHELP/requirements.txt
streamlit>=1.22.0
langchain>=0.0.267
langchain-community>=0.0.1
langchain-openai>=0.0.1
langchain-core>=0.1.7
openai>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pillow>=9.5.0
python-dotenv>=1.0.0
pypdf2>=3.0.1
chromadb>=0.4.6
pydantic>=1.10.8
